% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_5.R
\name{team_5}
\alias{team_5}
\title{Function from Team 5}
\usage{
team_5(file, tolerance = 0.1)
}
\arguments{
\item{file}{is the location of the data}

\item{tolerance}{is the value used for thinning the polygon}
}
\value{
ozplus is a data frame about how to plot the map of Australia
examples
file=system.file("gadm36_AUS_shp/gadm36_AUS_1.shp", package="Team5")
tolerance <- 0.1
team_5(file="gadm36_AUS_1.shp",0.01)
team_5(file,0.1) %>% ggplot(aes(x = long, y = lat, group = group)) + geom_polygon(fill="white",color="black",lwd=1)
}
\description{
Function from Team 5
}
