% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_6.R
\name{team_6}
\alias{team_6}
\title{solution of Lab 2 of team 6.}
\usage{
team_6(file, tolerance)
}
\arguments{
\item{file}{A path having shape file of selected country.}

\item{tolerance}{A number for tolerence.}
}
\value{
The dataframe
}
\description{
solution of Lab 2 of team 6.
}
\examples{
file<-"./data/gadm36_AUS_shp/gadm36_AUS_1.shp"
tolerance <- 0.1
team_6(file,tolerance) \%>\% ggplot(aes(x = long, y = lat, group = group)) + geom_polygon(fill="white",color="black",lwd=1)
}
